fix(multiclass): widen maxClasses to Long for JS-sentinel decoding
`MulticlassConfig.maxClasses` was typed `Int` (default `Int.MAX_VALUE`).
Production's `application-local.yaml` carried `maxClasses: 9007199254740991`
— `Number.MAX_SAFE_INTEGER`, the JavaScript "unlimited" sentinel from the
config-authoring tooling. That overflows Int, so Hoplite refused to decode
it and the server crash-looped on boot.

Widen the field (and the matching `multiclassMaxClasses` field on the
`Trainer.List` GMCP payload) to `Long`. JS clients already type it as
`number` and the actual sentinel value fits exactly in `Number`, so the
wire format is unchanged. Comparisons in `TrainerHandler` get an explicit
`.toLong()` on the player's unlocked-count.

Regression test in `AppConfigLoaderTest` decodes the JS sentinel via
Hoplite and asserts the field round-trips.

+++ src/main/kotlin/dev/ambon/config/AppConfig.kt
@@ -2350,8 +2350,11 @@ data class MulticlassConfig(
      * Maximum number of classes a player may have unlocked (including their starter class).
      * Defaults to effectively unlimited so existing deployments keep working unchanged;
      * curated `application.yaml` ships a tighter cap.
+     *
+     * Typed as [Long] so config overlays written by JavaScript tooling (which often uses
+     * `Number.MAX_SAFE_INTEGER` = 2^53−1 as an "unlimited" sentinel) decode without overflow.
      */
-    val maxClasses: Int = Int.MAX_VALUE,
+    val maxClasses: Long = Long.MAX_VALUE,
     /**
      * Exponential multiplier applied per additional class beyond the first trainer unlock.
      * Cost for the Nth trainer unlock is `goldCost * goldCostMultiplier^(N-1)`, so the first
